Move kms_keyring.h under the cpp/ subdirectory
Closes: #180
