+++ aws-encryption-sdk-cpp/include/aws/cryptosdk/cpp/kms_keyring.h
